requirements.txt
networkx
numpy
scikit-learn
sentence-transformers
pdfplumber
google-generativeai
python-dotenv
streamlit
watchdog
